+++ tests/fixtures/timer-apis-g45@test/extension.js
@@ -1,0 +1,13 @@
+import {Extension} from 'resource:///org/gnome/shell/extensions/extension.js';
+
+export default class WebApisGnome45Extension extends Extension {
+    enable() {
+        this._timer = setTimeout(() => {}, 100);
+        this._interval = setInterval(() => {}, 1000);
+    }
+
+    disable() {
+        clearTimeout(this._timer);
+        clearInterval(this._interval);
+    }
+}